Moved NameContext to Context.ValueContext
